weather-tracker/weather-backend/requirements.txt
flask==3.0.2
werkzeug==3.0.1
requests==2.31.0
prometheus-client==0.19.0
flask-cors==4.0.0
fastapi==0.110.0
python-dotenv==1.0.1
firebase-admin==6.4.0
uvicorn==0.27.1 